+++ Changelog.md
@@ -3,7 +3,8 @@
 ### 3.2.2:
 - Remove package legacy factories.
 - Adopt laravel 8 factories style.
-- Create unit tests for all models
+- Create unit tests for all models.
+- Create Unit Tests for all factories.
 
 ### 3.2.1:
 - Notify about active sessions passing the hours threshold.